This adds an index to the versions table to speed up the query in
Dependencies#show. The index is on the rubygem_id, position, and created_at columns, and is ordered by position ascending and created_at descending.
The index is only applied to rows where indexed = true.

The `order by` clause is used to ensure that an index scan is all that is needed to
run the query.

Before:

```
Gather Merge  (cost=48301.05..49691.58 rows=11918 width=878) (actual time=27.670..32.615 rows=16297 loops=1)
  Workers Planned: 2
  Workers Launched: 2
  ->  Sort  (cost=47301.02..47315.92 rows=5959 width=878) (actual time=17.273..17.662 rows=5432 loops=3)
        Sort Key: "position", created_at DESC
        Sort Method: external merge  Disk: 3624kB
        Worker 0:  Sort Method: quicksort  Memory: 2490kB
        Worker 1:  Sort Method: quicksort  Memory: 2431kB
        ->  Parallel Bitmap Heap Scan on versions  (cost=406.23..44624.37 rows=5959 width=878) (actual time=1.631..10.145 rows=5432 loops=3)
              Recheck Cond: (rubygem_id = ANY ('{15163,112014,26159,16914,15076,32035,173920,186981,183849,183850}'::integer[]))
              Filter: indexed
              Rows Removed by Filter: 17
              Heap Blocks: exact=3925
              ->  Bitmap Index Scan on index_versions_on_rubygem_id_and_number_and_platform  (cost=0.00..402.62 rows=15777 width=0) (actual time=3.251..3.251 rows=16347 loops=1)
                    Index Cond: (rubygem_id = ANY ('{15163,112014,26159,16914,15076,32035,173920,186981,183849,183850}'::integer[]))
Planning Time: 0.212 ms
Execution Time: 33.862 ms
```

After:

```
Index Only Scan using index_versions_on_rubygem_id_and_position_and_created_at on versions  (cost=0.43..613.35 rows=13564 width=28) (actual time=0.035..4.553 rows=16297 loops=1)
  Index Cond: (rubygem_id = ANY ('{15163,112014,26159,16914,15076,32035,173920,186981,183849,183850}'::integer[]))
  Heap Fetches: 0
Planning Time: 0.289 ms
Execution Time: 5.351 ms
```

This should help resolve the DOS issues we had last month!

Signed-off-by: Samuel Giddins <segiddins@segiddins.me>